# supabase_openapi.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "title": "standard public schema",
    "version": "14.5"
  },
  "host": "fwrkrpmqmqxlgvyzrizq.supabase.co:443",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json",
    "application/vnd.pgrst.object+json;nulls=stripped",
    "application/vnd.pgrst.object+json",
    "text/csv"
  ],
  "produces": [
    "application/json",
    "application/vnd.pgrst.object+json;nulls=stripped",
    "application/vnd.pgrst.object+json",
    "text/csv"
  ],
  "paths": {
...
    "/notifications": {
      "get": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.notifications.id"
          },
          {
            "$ref": "#/parameters/rowFilter.notifications.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.notifications.patient_uid"
          },
          {
            "$ref": "#/parameters/rowFilter.notifications.icon"
          },
          {
            "$ref": "#/parameters/rowFilter.notifications.message"
          },
          {
            "$ref": "#/parameters/rowFilter.notifications.is_read"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/order"
          },
          {
            "$ref": "#/parameters/range"
          },
          {
            "$ref": "#/parameters/rangeUnit"
          },
          {
            "$ref": "#/parameters/offset"
          },
          {
            "$ref": "#/parameters/limit"
          },
          {
            "$ref": "#/parameters/preferCount"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "items": {
                "$ref": "#/definitions/notifications"
              },
              "type": "array"
            }
          },
          "206": {
            "description": "Partial Content"
          }
        },
        "tags": [
          "notifications"
        ]
      },
      "post": {
        "parameters": [
          {
            "$ref": "#/parameters/body.notifications"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/preferPost"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "tags": [
          "notifications"
        ]
      },
      "delete": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.notifications.id"
          },
          {
            "$ref": "#/parameters/rowFilter.notifications.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.notifications.patient_uid"
          },
          {
            "$ref": "#/parameters/rowFilter.notifications.icon"
          },
          {
            "$ref": "#/parameters/rowFilter.notifications.message"
          },
          {
            "$ref": "#/parameters/rowFilter.notifications.is_read"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "notifications"
        ]
      },
      "patch": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.notifications.id"
          },
          {
            "$ref": "#/parameters/rowFilter.notifications.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.notifications.patient_uid"
          },
          {
            "$ref": "#/parameters/rowFilter.notifications.icon"
          },
          {
            "$ref": "#/parameters/rowFilter.notifications.message"
          },
          {
            "$ref": "#/parameters/rowFilter.notifications.is_read"
          },
          {
            "$ref": "#/parameters/body.notifications"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "notifications"
        ]
      }
    },
...
    "/profiles": {
      "get": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.profiles.id"
          },
          {
            "$ref": "#/parameters/rowFilter.profiles.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.profiles.patient_uid"
          },
          {
            "$ref": "#/parameters/rowFilter.profiles.parent_name"
          },
          {
            "$ref": "#/parameters/rowFilter.profiles.child_name"
          },
          {
            "$ref": "#/parameters/rowFilter.profiles.doctor_uid"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/order"
          },
          {
            "$ref": "#/parameters/range"
          },
          {
            "$ref": "#/parameters/rangeUnit"
          },
          {
            "$ref": "#/parameters/offset"
          },
          {
            "$ref": "#/parameters/limit"
          },
          {
            "$ref": "#/parameters/preferCount"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "items": {
                "$ref": "#/definitions/profiles"
              },
              "type": "array"
            }
          },
          "206": {
            "description": "Partial Content"
          }
        },
        "tags": [
          "profiles"
        ]
      },
      "post": {
        "parameters": [
          {
            "$ref": "#/parameters/body.profiles"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/preferPost"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "tags": [
          "profiles"
        ]
      },
      "delete": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.profiles.id"
          },
          {
            "$ref": "#/parameters/rowFilter.profiles.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.profiles.patient_uid"
          },
          {
            "$ref": "#/parameters/rowFilter.profiles.parent_name"
          },
          {
            "$ref": "#/parameters/rowFilter.profiles.child_name"
          },
          {
            "$ref": "#/parameters/rowFilter.profiles.doctor_uid"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "profiles"
        ]
      },
      "patch": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.profiles.id"
          },
          {
            "$ref": "#/parameters/rowFilter.profiles.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.profiles.patient_uid"
          },
          {
            "$ref": "#/parameters/rowFilter.profiles.parent_name"
          },
          {
            "$ref": "#/parameters/rowFilter.profiles.child_name"
          },
          {
            "$ref": "#/parameters/rowFilter.profiles.doctor_uid"
          },
          {
            "$ref": "#/parameters/body.profiles"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "profiles"
        ]
      }
    },
...
    "/patients": {
      "get": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.patients.id"
          },
          {
            "$ref": "#/parameters/rowFilter.patients.doctor_id"
          },
          {
            "$ref": "#/parameters/rowFilter.patients.name"
          },
          {
            "$ref": "#/parameters/rowFilter.patients.age"
          },
          {
            "$ref": "#/parameters/rowFilter.patients.condition"
          },
          {
            "$ref": "#/parameters/rowFilter.patients.phone"
          },
          {
            "$ref": "#/parameters/rowFilter.patients.parent_email"
          },
          {
            "$ref": "#/parameters/rowFilter.patients.notes"
          },
          {
            "$ref": "#/parameters/rowFilter.patients.practice_code"
          },
          {
            "$ref": "#/parameters/rowFilter.patients.patient_uid"
          },
          {
            "$ref": "#/parameters/rowFilter.patients.accuracy"
          },
          {
            "$ref": "#/parameters/rowFilter.patients.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.patients.updated_at"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/order"
          },
          {
            "$ref": "#/parameters/range"
          },
          {
            "$ref": "#/parameters/rangeUnit"
          },
          {
            "$ref": "#/parameters/offset"
          },
          {
            "$ref": "#/parameters/limit"
          },
          {
            "$ref": "#/parameters/preferCount"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "items": {
                "$ref": "#/definitions/patients"
              },
              "type": "array"
            }
          },
          "206": {
            "description": "Partial Content"
          }
        },
        "tags": [
          "patients"
        ]
      },
      "post": {
        "parameters": [
          {
            "$ref": "#/parameters/body.patients"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/preferPost"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "tags": [
          "patients"
        ]
      },
      "delete": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.patients.id"
          },
          {
            "$ref": "#/parameters/rowFilter.patients.doctor_id"
          },
          {
            "$ref": "#/parameters/rowFilter.patients.name"
          },
          {
            "$ref": "#/parameters/rowFilter.patients.age"
          },
          {
            "$ref": "#/parameters/rowFilter.patients.condition"
          },
          {
            "$ref": "#/parameters/rowFilter.patients.phone"
          },
          {
            "$ref": "#/parameters/rowFilter.patients.parent_email"
          },
          {
            "$ref": "#/parameters/rowFilter.patients.notes"
          },
          {
            "$ref": "#/parameters/rowFilter.patients.practice_code"
          },
          {
            "$ref": "#/parameters/rowFilter.patients.patient_uid"
          },
          {
            "$ref": "#/parameters/rowFilter.patients.accuracy"
          },
          {
            "$ref": "#/parameters/rowFilter.patients.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.patients.updated_at"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "patients"
        ]
      },
      "patch": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.patients.id"
          },
          {
            "$ref": "#/parameters/rowFilter.patients.doctor_id"
          },
          {
            "$ref": "#/parameters/rowFilter.patients.name"
          },
          {
            "$ref": "#/parameters/rowFilter.patients.age"
          },
          {
            "$ref": "#/parameters/rowFilter.patients.condition"
          },
          {
            "$ref": "#/parameters/rowFilter.patients.phone"
          },
          {
            "$ref": "#/parameters/rowFilter.patients.parent_email"
          },
          {
            "$ref": "#/parameters/rowFilter.patients.notes"
          },
          {
            "$ref": "#/parameters/rowFilter.patients.practice_code"
          },
          {
            "$ref": "#/parameters/rowFilter.patients.patient_uid"
          },
          {
            "$ref": "#/parameters/rowFilter.patients.accuracy"
          },
          {
            "$ref": "#/parameters/rowFilter.patients.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.patients.updated_at"
          },
          {
            "$ref": "#/parameters/body.patients"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "patients"
        ]
      }
    },
...
  },
  "definitions": {
...
    "notifications": {
      "required": [
        "id"
      ],
      "properties": {
        "id": {
          "description": "Note:\nThis is a Primary Key.<pk/>",
          "format": "bigint",
          "type": "integer"
        },
        "created_at": {
          "default": "now()",
          "format": "timestamp with time zone",
          "type": "string"
        },
        "patient_uid": {
          "format": "text",
          "type": "string"
        },
        "icon": {
          "format": "text",
          "type": "string"
        },
        "message": {
          "format": "text",
          "type": "string"
        },
        "is_read": {
          "default": false,
          "format": "boolean",
          "type": "boolean"
        }
      },
      "type": "object"
    },
...
    "profiles": {
      "required": [
        "id"
      ],
      "properties": {
        "id": {
          "description": "Note:\nThis is a Primary Key.<pk/>",
          "format": "bigint",
          "type": "integer"
        },
        "created_at": {
          "default": "now()",
          "format": "timestamp with time zone",
          "type": "string"
        },
        "patient_uid": {
          "format": "text",
          "type": "string"
        },
        "parent_name": {
          "format": "text",
          "type": "string"
        },
        "child_name": {
          "format": "text",
          "type": "string"
        },
        "doctor_uid": {
          "format": "text",
          "type": "string"
        }
      },
      "type": "object"
    },
...
    "patients": {
      "required": [
        "id",
        "doctor_id",
        "name",
        "practice_code",
        "created_at",
        "updated_at"
      ],
      "properties": {
        "id": {
          "default": "gen_random_uuid()",
          "description": "Note:\nThis is a Primary Key.<pk/>",
          "format": "uuid",
          "type": "string"
        },
        "doctor_id": {
          "description": "Note:\nThis is a Foreign Key to `therapists.id`.<fk table='therapists' column='id'/>",
          "format": "text",
          "type": "string"
        },
        "name": {
          "format": "text",
          "type": "string"
        },
        "age": {
          "format": "integer",
          "type": "integer"
        },
        "condition": {
          "format": "text",
          "type": "string"
        },
        "phone": {
          "format": "text",
          "type": "string"
        },
        "parent_email": {
          "format": "text",
          "type": "string"
        },
        "notes": {
          "format": "text",
          "type": "string"
        },
        "practice_code": {
          "format": "text",
          "type": "string"
        },
        "patient_uid": {
          "format": "text",
          "type": "string"
        },
        "accuracy": {
          "format": "integer",
          "type": "integer"
        },
        "created_at": {
          "default": "now()",
          "format": "timestamp with time zone",
          "type": "string"
        },
        "updated_at": {
          "default": "now()",
          "format": "timestamp with time zone",
          "type": "string"
        }
      },
      "type": "object"
    },
...
  },
  "parameters": {
...
    "preferReturn": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "return=representation",
        "return=minimal",
        "return=none"
      ],
      "in": "header",
      "type": "string"
    },
    "preferCount": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "count=none"
      ],
      "in": "header",
      "type": "string"
    },
    "preferPost": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "return=representation",
        "return=minimal",
        "return=none",
        "resolution=ignore-duplicates",
        "resolution=merge-duplicates"
      ],
      "in": "header",
      "type": "string"
    },
    "select": {
      "name": "select",
      "description": "Filtering Columns",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
    "order": {
      "name": "order",
      "description": "Ordering",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "range": {
      "name": "Range",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "header",
      "type": "string"
    },
    "rangeUnit": {
      "name": "Range-Unit",
      "description": "Limiting and Pagination",
      "required": false,
      "default": "items",
      "in": "header",
      "type": "string"
    },
    "offset": {
      "name": "offset",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "limit": {
      "name": "limit",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
    "body.notifications": {
      "name": "notifications",
      "description": "notifications",
      "required": false,
      "in": "body",
      "schema": {
        "$ref": "#/definitions/notifications"
      }
    },
    "rowFilter.notifications.id": {
      "name": "id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.notifications.created_at": {
      "name": "created_at",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.notifications.patient_uid": {
      "name": "patient_uid",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.notifications.icon": {
      "name": "icon",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.notifications.message": {
      "name": "message",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.notifications.is_read": {
      "name": "is_read",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
    "body.profiles": {
      "name": "profiles",
      "description": "profiles",
      "required": false,
      "in": "body",
      "schema": {
        "$ref": "#/definitions/profiles"
      }
    },
    "rowFilter.profiles.id": {
      "name": "id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.profiles.created_at": {
      "name": "created_at",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.profiles.patient_uid": {
      "name": "patient_uid",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.profiles.parent_name": {
      "name": "parent_name",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.profiles.child_name": {
      "name": "child_name",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.profiles.doctor_uid": {
      "name": "doctor_uid",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
    "body.patients": {
      "name": "patients",
      "description": "patients",
      "required": false,
      "in": "body",
      "schema": {
        "$ref": "#/definitions/patients"
      }
    },
    "rowFilter.patients.id": {
      "name": "id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.patients.doctor_id": {
      "name": "doctor_id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.patients.name": {
      "name": "name",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.patients.age": {
      "name": "age",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.patients.condition": {
      "name": "condition",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.patients.phone": {
      "name": "phone",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.patients.parent_email": {
      "name": "parent_email",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.patients.notes": {
      "name": "notes",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.patients.practice_code": {
      "name": "practice_code",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.patients.patient_uid": {
      "name": "patient_uid",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.patients.accuracy": {
      "name": "accuracy",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.patients.created_at": {
      "name": "created_at",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.patients.updated_at": {
      "name": "updated_at",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
  },
...
}